little fix

+++ LobsterGameEngine/src/graphics/meshes/MeshFactory.h
@@ -10,7 +10,7 @@ namespace Lobster
     public:
         static VertexArray* Cube();
 		static VertexArray* Plane();
-		static VertexArray* Sphere(int radius, int du, int dv);
+		static VertexArray* Sphere(float radius, int du, int dv);
 		static VertexArray* Grid(int row, int col);
     };
